# docs/swagger.yaml
basePath: /
definitions:
  model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: API для управления подписками и получения агрегированных сумм.
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку
      parameters:
      - description: Данные подписки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID. Требует query-параметр user_id
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет существующую подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные подписки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Возвращает сумму цен подписок пользователя, пересекающих интервал
        [from; to]
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: Имя сервиса
        in: query
        name: service_name
        type: string
      - description: Дата начала (RFC3339 или YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Дата конца (RFC3339 или YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Сумма по подпискам за период
      tags:
      - subscriptions
swagger: "2.0"